API/requirements.txt
spacy==2.3.2
spacy-lookups-data==0.3.2
# torch with cuda 10.2 - for other cuda versions or no cuda go to https://pytorch.org/get-started/locally/
torch==1.5.1
torchvision==0.6.1
numpy==1.19.1
tqdm==4.48.2
matplotlib==3.3.1
pandas==1.1.1
Pillow==7.2.0
opencv-python==4.4.0.42
tqdm==4.48.2
streamlit~=0.66.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz#egg=en_core_web_sm

